OGC Refactoring Part 10 - Establish features package structure (similar to /igeserver/profiles) - Create features package - Move OGC package to features - Remove spring profile annotations form ogc files - Create OgcConfig and add spring profile 'ogc-api' - Create FeatureFilter and add it to SpringConfiguration to exclude all packages in /igeserver/features - Rename package 'ogc' to 'ogcApi'
