+++ server/src/main/java/de/ingrid/igeserver/features/ogcApi/distributionHelper/DistributionTypeInfo.kt
@@ -1,0 +1,30 @@
+/**
+ * ==================================================
+ * Copyright (C) 2024 wemove digital solutions GmbH
+ * ==================================================
+ * Licensed under the EUPL, Version 1.2 or – as soon they will be
+ * approved by the European Commission - subsequent versions of the
+ * EUPL (the "Licence");
+ *
+ * You may not use this work except in compliance with the Licence.
+ * You may obtain a copy of the Licence at:
+ *
+ * https://joinup.ec.europa.eu/software/page/eupl
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the Licence is distributed on an "AS IS" basis,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the Licence for the specific language governing permissions and
+ * limitations under the Licence.
+ */
+package de.ingrid.igeserver.features.ogcApi.distributionHelper
+
+
+/**
+ * @param id the ID of the distribution helper
+ * @param name the display name of the distribution helper
+ * @param description a description for the distribution helper
+ * @param profiles in which profiles can this distribution helper be used
+ *
+ */
+data class DistributionTypeInfo(val id: String, val name: String, val description: String, val profiles: List<String>)
