+++ server/src/main/java/de/ingrid/mdek/upload/configuration/FeatureFilter.java
@@ -1,0 +1,42 @@
+/**
+ * ==================================================
+ * Copyright (C) 2023-2024 wemove digital solutions GmbH
+ * ==================================================
+ * Licensed under the EUPL, Version 1.2 or – as soon they will be
+ * approved by the European Commission - subsequent versions of the
+ * EUPL (the "Licence");
+ *
+ * You may not use this work except in compliance with the Licence.
+ * You may obtain a copy of the Licence at:
+ *
+ * https://joinup.ec.europa.eu/software/page/eupl
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the Licence is distributed on an "AS IS" basis,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the Licence for the specific language governing permissions and
+ * limitations under the Licence.
+ */
+package de.ingrid.mdek.upload.configuration;
+
+import org.springframework.core.type.ClassMetadata;
+import org.springframework.core.type.classreading.MetadataReader;
+import org.springframework.core.type.classreading.MetadataReaderFactory;
+import org.springframework.core.type.filter.TypeFilter;
+
+import java.io.IOException;
+
+/**
+ * This filter allows all classes to be instantiated directly beneath package "de.ingrid.igeserver.feature"
+ * but not the sub-packages. These need to be included by a feature configuration, which is activated by a spring profile
+ */
+public class FeatureFilter implements TypeFilter {
+
+    @Override
+    public boolean match(MetadataReader metadataReader,
+                         MetadataReaderFactory metadataReaderFactory) throws IOException {
+        ClassMetadata classMetadata = metadataReader.getClassMetadata();
+        String fullyQualifiedName = classMetadata.getClassName();
+        return fullyQualifiedName.startsWith("de.ingrid.igeserver.features.") && fullyQualifiedName.lastIndexOf('.') != "de.ingrid.igeserver.features.".length() - 1;
+    }
+}
